Skip deletion of  preserved machines with unregistered nodes (#425)

+++ cluster-autoscaler/cloudprovider/mcm/mcm_cloud_provider.go
@@ -23,6 +23,7 @@ package mcm
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"slices"
 	"strconv"
@@ -41,7 +42,7 @@ import (
 	"k8s.io/autoscaler/cluster-autoscaler/cloudprovider"
 	"k8s.io/autoscaler/cluster-autoscaler/config"
 	"k8s.io/autoscaler/cluster-autoscaler/simulator/framework"
-	"k8s.io/autoscaler/cluster-autoscaler/utils/errors"
+	caerrors "k8s.io/autoscaler/cluster-autoscaler/utils/errors"
 	"k8s.io/klog/v2"
 )
 
@@ -68,6 +69,10 @@ const (
 	// MaxNodeProvisionTimeAnnotation is the annotation key for the value of NodeGroupAutoscalingOptions.MaxNodeProvisionTime
 	MaxNodeProvisionTimeAnnotation = "autoscaler.gardener.cloud/max-node-provision-time"
 )
+
+// ErrNoNodesToDelete is a sentinel error that indicates that nodes could not be deleted either due to machine being in Failed or Terminating phase,
+// because machine was preserved or because the node was annotated with scale-down-disabled=true
+var ErrNoNodesToDelete = errors.New("no nodes deleted: all candidates skipped due to termination, preservation, or scale-down-disabled annotation")
 
 // MCMCloudProvider implements the cloud provider interface for machine-controller-manager
 // Reference: https://github.com/gardener/machine-controller-manager
@@ -170,7 +175,7 @@ func (mcm *mcmCloudProvider) HasInstance(*apiv1.Node) (bool, error) {
 }
 
 // Pricing returns pricing model for this cloud provider or error if not available.
-func (mcm *mcmCloudProvider) Pricing() (cloudprovider.PricingModel, errors.AutoscalerError) {
+func (mcm *mcmCloudProvider) Pricing() (cloudprovider.PricingModel, caerrors.AutoscalerError) {
 	return nil, cloudprovider.ErrNotImplemented
 }
 
@@ -401,6 +406,9 @@ func (ngImpl *nodeGroup) DeleteNodes(nodes []*apiv1.Node) error {
 // ForceDeleteNodes deletes the nodes from the group without checking for minSize constraint.
 // It's expected that the method won't be called for nodes that aren't part of ANY machine deployment.
 func (ngImpl *nodeGroup) ForceDeleteNodes(nodes []*apiv1.Node) error {
+	if len(nodes) == 0 {
+		return nil
+	}
 	var toBeDeletedMachineInfos = make([]machineInfo, 0, len(nodes))
 	for _, node := range nodes {
 		belongs, mInfo, err := ngImpl.belongs(node)
@@ -409,15 +417,22 @@ func (ngImpl *nodeGroup) ForceDeleteNodes(nodes []*apiv1.Node) error {
 		} else if !belongs {
 			return fmt.Errorf("%s belongs to a different MachineDeployment than %q", node.Name, ngImpl.Name)
 		}
-		if mInfo.FailedOrTerminating {
+		// Failed machines that are preserved are covered by the first case.
+		// If the machine is not preserved, but is in failed state, the second case is executed.
+		// The ordering of the cases needs to be maintained for accurate logging.
+		switch {
+		case mInfo.MachinePreserved:
+			klog.V(3).Infof("for NodeGroup %q, Node %q corresponding to Machine %q is marked as preserved - skipping deletion", ngImpl.Name, node.Name, mInfo.Key.Name)
+		case mInfo.FailedOrTerminating:
 			klog.V(3).Infof("for NodeGroup %q, Machine %q is already marked as terminating - skipping deletion", ngImpl.Name, mInfo.Key.Name)
-			continue
+		case eligibility.HasNoScaleDownAnnotation(node):
+			klog.V(3).Infof("for NodeGroup %q, Node %q corresponding to Machine %q is marked with ScaleDownDisabledAnnotation %q - skipping deletion", ngImpl.Name, node.Name, mInfo.Key.Name, eligibility.ScaleDownDisabledKey)
+		default:
+			toBeDeletedMachineInfos = append(toBeDeletedMachineInfos, *mInfo)
 		}
-		if eligibility.HasNoScaleDownAnnotation(node) {
-			klog.V(4).Infof("for NodeGroup %q, Node %q corresponding to Machine %q is marked with ScaleDownDisabledAnnotation %q - skipping deletion", ngImpl.Name, node.Name, mInfo.Key.Name, eligibility.ScaleDownDisabledKey)
-			continue
-		}
-		toBeDeletedMachineInfos = append(toBeDeletedMachineInfos, *mInfo)
+	}
+	if len(toBeDeletedMachineInfos) == 0 {
+		return ErrNoNodesToDelete
 	}
 	return ngImpl.deleteMachines(toBeDeletedMachineInfos)
 }
